Refactor voting.js into module - step #1
